+++ src/Enterspeed.Source.Sdk/Configuration/EnterspeedConfiguration.cs
@@ -8,9 +8,9 @@
         public string ApiKey { get; set; }
 
         /// <summary>
-        /// Gets or sets the base URL.
+        /// Gets or sets the base URL. Default: https://api.enterspeed.com.
         /// </summary>
-        public string BaseUrl { get; set; }
+        public string BaseUrl { get; set; } = "https://api.enterspeed.com";
 
         /// <summary>
         /// Gets or sets timeout in seconds. Default: 60 seconds.
@@ -20,6 +20,11 @@
         /// <summary>
         /// Gets the current version for the Ingest Endpoint.
         /// </summary>
-        public string IngestVersion => "1";
+        public string IngestVersion => "2";
+
+        /// <summary>
+        /// Gets or sets the SystemInformation.
+        /// </summary>
+        public string SystemInformation { get; set; }
     }
 }